# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "欺骗防御接口文档",
        "contact": {
            "name": "akita.tian",
            "email": "chronoyq@163.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8082",
    "paths": {
        "/api/public/attack/falco": {
            "post": {
                "description": "创建falco异常事件",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "非前端接口"
                ],
                "summary": "创建falco异常事件",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/public/attack/protocol": {
            "post": {
                "description": "创建协议代理攻击事件",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "非前端接口"
                ],
                "summary": "创建协议代理攻击事件",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/attack/display/attackIP": {
            "get": {
                "description": "攻击IP统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "大屏展示"
                ],
                "summary": "攻击IP统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.AttackStatistics"
                        }
                    }
                }
            }
        },
        "/api/v1/attack/display/location": {
            "get": {
                "description": "攻击位置统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "大屏展示"
                ],
                "summary": "攻击位置统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.AttackStatistics"
                        }
                    }
                }
            }
        },
        "/api/v1/attack/display/probeIP": {
            "get": {
                "description": "攻击业务IP统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "大屏展示"
                ],
                "summary": "攻击业务IP统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.AttackStatistics"
                        }
                    }
                }
            }
        },
        "/api/v1/attack/display/protocol": {
            "get": {
                "description": "攻击服务类型统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "大屏展示"
                ],
                "summary": "攻击服务类型统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.AttackStatistics"
                        }
                    }
                }
            }
        },
        "/api/v1/attack/falco": {
            "post": {
                "description": "falco攻击",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "威胁感知"
                ],
                "summary": "falco攻击",
                "parameters": [
                    {
                        "description": "StartTime",
                        "name": "StartTime",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.FalcoEventSelectPayload"
                        }
                    },
                    {
                        "description": "EndTime",
                        "name": "EndTime",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.FalcoEventSelectPayload"
                        }
                    },
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.FalcoEventSelectPayload"
                        }
                    },
                    {
                        "description": "PageNumber",
                        "name": "PageNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.FalcoEventSelectPayload"
                        }
                    },
                    {
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.FalcoEventSelectPayload"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.FalcoSelectResultPayload"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/attack/falco/:id": {
            "get": {
                "description": "falco攻击详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "威胁感知"
                ],
                "summary": "falco攻击详情",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "int valid",
                        "name": "int",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OutputFields"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/attack/token": {
            "post": {
                "description": "密签跟踪日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "威胁感知"
                ],
                "summary": "密签跟踪日志",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectPayload"
                        }
                    },
                    {
                        "description": "StartTime",
                        "name": "StartTime",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectPayload"
                        }
                    },
                    {
                        "description": "EndTime",
                        "name": "EndTime",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectPayload"
                        }
                    },
                    {
                        "description": "AttackIP",
                        "name": "AttackIP",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectPayload"
                        }
                    },
                    {
                        "description": "ServerType",
                        "name": "ServerType",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectPayload"
                        }
                    },
                    {
                        "description": "PageNumber",
                        "name": "PageNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectPayload"
                        }
                    },
                    {
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectPayload"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comm.TokenTraceSelectResultPayload"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "comm.AttackStatistics": {
            "type": "object",
            "properties": {
                "Count": {
                    "description": "出现次数",
                    "type": "integer"
                },
                "Data": {
                    "description": "攻击信息",
                    "type": "string"
                }
            }
        },
...
        "comm.FalcoEventSelectPayload": {
            "type": "object",
            "required": [
                "PageNumber",
                "PageSize"
            ],
            "properties": {
                "EndTime": {
                    "description": "结束时间",
                    "type": "string"
                },
                "PageNumber": {
                    "description": "页number",
                    "type": "integer"
                },
                "PageSize": {
                    "description": "页size",
                    "type": "integer"
                },
                "Payload": {
                    "description": "查找payload",
                    "type": "string"
                },
                "StartTime": {
                    "description": "开始时间",
                    "type": "string"
                }
            }
        },
        "comm.FalcoSelectResultPayload": {
            "type": "object",
            "properties": {
                "DownloadPath": {
                    "description": "文件下载链接",
                    "type": "string"
                },
                "Event": {
                    "description": "falco事件",
                    "type": "string"
                },
                "FileFlag": {
                    "description": "文件标记",
                    "type": "boolean"
                },
                "HoneypotName": {
                    "description": "蜜罐名称",
                    "type": "string"
                },
                "ID": {
                    "description": "falco攻击日志ID",
                    "type": "integer"
                },
                "Level": {
                    "description": "事件等级",
                    "type": "string"
                },
                "Output": {
                    "description": "输出",
                    "type": "string"
                },
                "Time": {
                    "description": "发生时间",
                    "type": "string"
                }
            }
        },
...
        "models.OutputFields": {
            "type": "object",
            "properties": {
                "container.id": {
                    "description": "容器ID",
                    "type": "string"
                },
                "container.image.repository": {
                    "description": "镜像仓库",
                    "type": "string"
                },
                "container.name": {
                    "description": "容器名称",
                    "type": "string"
                },
                "evt.time": {
                    "description": "事件时间",
                    "type": "integer"
                },
                "fd.name": {
                    "description": "上传的文件名称",
                    "type": "string"
                },
                "k8s.ns.name": {
                    "description": "命名空间",
                    "type": "string"
                },
                "k8s.pod.name": {
                    "description": "pod名称",
                    "type": "string"
                },
                "proc.cmdline": {
                    "description": "命令行",
                    "type": "string"
                },
                "proc.name": {
                    "description": "进程名称",
                    "type": "string"
                },
                "proc.pname": {
                    "description": "父进程名称",
                    "type": "string"
                },
                "proc.tty": {
                    "description": "TTY",
                    "type": "integer"
                },
                "user.loginuid": {
                    "description": "登录UID",
                    "type": "integer"
                },
                "user.name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
    }
}